python/requirements.txt
mypy

# Main libs
python-dotenv

beautifulsoup4
colorama
requests
redis

# Type stubs
types-beautifulsoup4
types-colorama
types-requests
types-redis